Rework JUnit integration to work from TestSuite
The original Nachos JUnit integration taht this repo was based on did not work
with defining multiple classes within which to do the testing. The major
limitation was that each class had to be derived from 'TestHarness', which
would spawn up a new instance of Nachos to run the tests contained in that
class.  Spawning up a whole bunch of Nachos instances was neither working very
well, nor the desired behaviour.

This patch updates the Nachos JUnit integration to work with a JUnit Test
Suite, which only spins up one instance of Nachos, then runs all defined tests
across multiple classes within that one instance.

One caveat of using a test suite is that all of your Test Classes must be
expliclty listed so that the test suite can find them. This is accomplished in
the UnitTests.java file associated within each nachos.proj package.

Simply create new classes that follow the example layed out in
ExampleTests.java and add them to your UnitTests test suite to run them.

Running them is the same as before.  Simply choose the Run Configuration for
the project at hand and you are good to go.